requirements.txt
# Telegram
telethon>=1.34.0
cryptg>=0.4.0

# MT5
MetaTrader5>=5.0.45

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
aiofiles>=23.0.0

